+++ robot_task/include/robot_task/MultiGoalActionServer.h
@@ -42,6 +42,7 @@
 #ifndef ACTIONLIB_MULTIGOAL_ACTION_SERVER_H_
 #define ACTIONLIB_MULTIGOAL_ACTION_SERVER_H_
 
+#include <boost/date_time/posix_time/posix_time.hpp>
 #include <boost/thread/condition.hpp>
 #include <ros/ros.h>
 #include <actionlib/server/action_server.h>
@@ -420,7 +421,7 @@ namespace actionlib {
 							 (int)stopped_tasks, (int)started_tasks, (int)active_goals.size());
 				}
 				
-				execute_condition_.timed_wait(lock, boost::posix_time::milliseconds(loop_duration.toSec() * 1000.0f));
+        execute_condition_.timed_wait(lock, boost::posix_time::milliseconds((int)(loop_duration.toSec() * 1000.0f)));
 		}
     
 		while(new_goals.size()>0) new_goals.remove(new_goals.begin());
